requirements.txt
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
exceptiongroup==1.2.2
fonttools==4.53.1
frozendict==2.4.4
html5lib==1.1
idna==3.7
iniconfig==2.0.0
kiwisolver==1.4.5
lxml==5.3.0
matplotlib==3.9.2
multitasking==0.0.11
numpy==2.1.0
packaging==24.1
pandas==2.2.2
peewee==3.17.6
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
protobuf==5.29.0
pyparsing==3.1.2
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.6
tomli==2.0.1
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
yfinance==0.2.41
